Moved fallback detector to a distinct folder
