+++ src/universalDetectors/fallback.ts
